# docs/openapi.yaml
components:
  schemas:
...
    DeleteResponse:
      description: Generic delete response.
      properties:
        deleted:
          examples:
          - true
          title: Deleted
          type: boolean
      required:
      - deleted
      title: DeleteResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    NutrientInput:
      description: Optional nutrient input attached to a meal item or seed record.
      properties:
        amount:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 120
          title: Amount
        assumptions_json:
          items: {}
          title: Assumptions Json
          type: array
        confidence:
          anyOf:
          - maximum: 1.0
            minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 0.95
          title: Confidence
        evidence_json:
          additionalProperties: true
          title: Evidence Json
          type: object
        source_kind:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - manual
          title: Source Kind
        unit:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - mg
          title: Unit
      title: NutrientInput
      type: object
    PatternCreate:
      description: Create a reusable meal item pattern.
      examples:
      - aliases:
        - воппер
        - вопер
        - whopper
        default_carbs_g: 51
        default_fat_g: 35
        default_fiber_g: 3
        default_grams: 270
        default_kcal: 635
        default_protein_g: 28
        display_name: Whopper
        key: whopper
        prefix: bk
        source_url: https://origin.bk.com/pdfs/nutrition.pdf
      properties:
        aliases:
          examples:
          - - воппер
          items:
            type: string
          title: Aliases
          type: array
        default_carbs_g:
          default: 0
          examples:
          - 51
          minimum: 0.0
          title: Default Carbs G
          type: number
        default_fat_g:
          default: 0
          examples:
          - 35
          minimum: 0.0
          title: Default Fat G
          type: number
        default_fiber_g:
          default: 0
          examples:
          - 3
          minimum: 0.0
          title: Default Fiber G
          type: number
        default_grams:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 270
          title: Default Grams
        default_kcal:
          default: 0
          examples:
          - 635
          minimum: 0.0
          title: Default Kcal
          type: number
        default_protein_g:
          default: 0
          examples:
          - 28
          minimum: 0.0
          title: Default Protein G
          type: number
        display_name:
          examples:
          - Whopper
          title: Display Name
          type: string
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - https://example.test/item.png
          title: Image Url
        is_verified:
          default: false
          title: Is Verified
          type: boolean
        key:
          examples:
          - whopper
          title: Key
          type: string
        nutrients_json:
          additionalProperties:
            anyOf:
            - $ref: '#/components/schemas/NutrientInput'
            - type: number
            - type: 'null'
          examples:
          - sodium_mg:
              amount: 980
              unit: mg
          title: Nutrients Json
          type: object
        per_100g_carbs_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 19
          title: Per 100G Carbs G
        per_100g_fat_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 16
          title: Per 100G Fat G
        per_100g_kcal:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 260
          title: Per 100G Kcal
        per_100g_protein_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 10
          title: Per 100G Protein G
        prefix:
          examples:
          - bk
          title: Prefix
          type: string
        source_confidence:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - official_pdf
          title: Source Confidence
        source_file:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - bk.generated.pdf
          title: Source File
        source_name:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - Burger King official PDF
          title: Source Name
        source_page:
          anyOf:
          - minimum: 1.0
            type: integer
          - type: 'null'
          examples:
          - 1
          title: Source Page
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - https://example.test
          title: Source Url
      required:
      - prefix
      - key
      - display_name
      title: PatternCreate
      type: object
    PatternPageResponse:
      description: Paginated pattern list response.
      properties:
        items:
          items:
            $ref: '#/components/schemas/PatternResponse'
          title: Items
          type: array
        limit:
          title: Limit
          type: integer
        offset:
          title: Offset
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - limit
      - offset
      title: PatternPageResponse
      type: object
    PatternPatch:
      description: Patch a reusable meal item pattern.
      properties:
        aliases:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Aliases
        default_carbs_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Default Carbs G
        default_fat_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Default Fat G
        default_fiber_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Default Fiber G
        default_grams:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Default Grams
        default_kcal:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Default Kcal
        default_protein_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Default Protein G
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        is_archived:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Archived
        is_verified:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Verified
        key:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
        nutrients_json:
          anyOf:
          - additionalProperties:
              anyOf:
              - $ref: '#/components/schemas/NutrientInput'
              - type: number
              - type: 'null'
            type: object
          - type: 'null'
          title: Nutrients Json
        per_100g_carbs_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Per 100G Carbs G
        per_100g_fat_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Per 100G Fat G
        per_100g_kcal:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Per 100G Kcal
        per_100g_protein_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Per 100G Protein G
        prefix:
          anyOf:
          - type: string
          - type: 'null'
          title: Prefix
        source_confidence:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Confidence
        source_file:
          anyOf:
          - type: string
          - type: 'null'
          title: Source File
        source_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Name
        source_page:
          anyOf:
          - minimum: 1.0
            type: integer
          - type: 'null'
          title: Source Page
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
      title: PatternPatch
      type: object
    PatternResponse:
      description: Pattern response.
      properties:
        aliases:
          items:
            type: string
          title: Aliases
          type: array
        created_at:
          format: date-time
          title: Created At
          type: string
        default_carbs_g:
          default: 0
          examples:
          - 51
          minimum: 0.0
          title: Default Carbs G
          type: number
        default_fat_g:
          default: 0
          examples:
          - 35
          minimum: 0.0
          title: Default Fat G
          type: number
        default_fiber_g:
          default: 0
          examples:
          - 3
          minimum: 0.0
          title: Default Fiber G
          type: number
        default_grams:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 270
          title: Default Grams
        default_kcal:
          default: 0
          examples:
          - 635
          minimum: 0.0
          title: Default Kcal
          type: number
        default_protein_g:
          default: 0
          examples:
          - 28
          minimum: 0.0
          title: Default Protein G
          type: number
        display_name:
          examples:
          - Whopper
          title: Display Name
          type: string
        id:
          format: uuid
          title: Id
          type: string
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - https://example.test/item.png
          title: Image Url
        is_archived:
          title: Is Archived
          type: boolean
        is_verified:
          default: false
          title: Is Verified
          type: boolean
        key:
          examples:
          - whopper
          title: Key
          type: string
        last_used_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Used At
        matched_alias:
          anyOf:
          - type: string
          - type: 'null'
          title: Matched Alias
        nutrients_json:
          additionalProperties:
            anyOf:
            - $ref: '#/components/schemas/NutrientInput'
            - type: number
            - type: 'null'
          examples:
          - sodium_mg:
              amount: 980
              unit: mg
          title: Nutrients Json
          type: object
        per_100g_carbs_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 19
          title: Per 100G Carbs G
        per_100g_fat_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 16
          title: Per 100G Fat G
        per_100g_kcal:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 260
          title: Per 100G Kcal
        per_100g_protein_g:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          examples:
          - 10
          title: Per 100G Protein G
        prefix:
          examples:
          - bk
          title: Prefix
          type: string
        source_confidence:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - official_pdf
          title: Source Confidence
        source_file:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - bk.generated.pdf
          title: Source File
        source_name:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - Burger King official PDF
          title: Source Name
        source_page:
          anyOf:
          - minimum: 1.0
            type: integer
          - type: 'null'
          examples:
          - 1
          title: Source Page
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          examples:
          - https://example.test
          title: Source Url
        updated_at:
          format: date-time
          title: Updated At
          type: string
        usage_count:
          title: Usage Count
          type: integer
      required:
      - prefix
      - key
      - display_name
      - id
      - usage_count
      - is_archived
      - created_at
      - updated_at
      title: PatternResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  summary: Informational food diary API.
  title: Glucotracker API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /patterns:
    get:
      description: List active patterns.
      operationId: listPatterns
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatternPageResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: List Patterns
      tags:
      - patterns
      x-glucotracker-scoped: true
    post:
      description: Create a reusable meal pattern.
      operationId: createPattern
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatternCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatternResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Create Pattern
      tags:
      - patterns
      x-glucotracker-scoped: true
  /patterns/search:
    get:
      description: Search active patterns by prefix, key, display name, and aliases.
      operationId: searchPatterns
      parameters:
      - in: query
        name: q
        required: true
        schema:
          title: Q
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PatternResponse'
                title: Response Searchpatterns
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Search Patterns
      tags:
      - patterns
      x-glucotracker-scoped: true
  /patterns/{pattern_id}:
    delete:
      description: Soft-delete a pattern by archiving it.
      operationId: deletePattern
      parameters:
      - in: path
        name: pattern_id
        required: true
        schema:
          format: uuid
          title: Pattern Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Delete Pattern
      tags:
      - patterns
      x-glucotracker-scoped: true
    get:
      description: Return a pattern by id.
      operationId: getPattern
      parameters:
      - in: path
        name: pattern_id
        required: true
        schema:
          format: uuid
          title: Pattern Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatternResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Pattern
      tags:
      - patterns
      x-glucotracker-scoped: true
    patch:
      description: Patch a reusable meal pattern.
      operationId: patchPattern
      parameters:
      - in: path
        name: pattern_id
        required: true
        schema:
          format: uuid
          title: Pattern Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatternPatch'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatternResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Patch Pattern
      tags:
      - patterns
      x-glucotracker-scoped: true
...